% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{learnr_downloads}
\alias{learnr_downloads}
\title{R Studio data for \code{learnr} downloads}
\format{
\subsection{\code{learnr_downloads}}{

A data frame with 7 observations of 2 variables.
\describe{
\item{year}{year of downloads}
\item{annual downloads}{the number of downloads of \code{learnr} for the respective
year}
}
}
}
\source{
\url{http://cran-logs.rstudio.com}
}
\usage{
learnr_downloads
}
\description{
A data downloaded with the \code{cranlogs} package from R Studio's CRAN mirror.
The subset contains data about the annual downloads of \code{learnr} package
since its release in 2017.
}
\keyword{datasets}
